# swagger.json
{
 "swagger": "2.0",
 "info": {
  "description": "Resource for managing Faros workspaces",
  "title": "Faros",
  "contact": {
   "name": "Mangirdas",
   "url": "https://faros.sh",
   "email": "mangirdas@judeikis.LT"
  },
  "license": {
   "name": "MIT",
   "url": "http://mit.org"
  },
  "version": "1.0.0"
 },
 "paths": {
...
  "/faros.sh/api/v1alpha1/organizations": {
   "get": {
    "consumes": [
     "application/json"
    ],
    "produces": [
     "application/json"
    ],
    "summary": "List organizations",
    "operationId": "listOrganizations",
    "responses": {
     "200": {
      "description": "OK",
      "schema": {
       "$ref": "#/definitions/v1alpha1.OrganizationList"
      }
     },
     "301": {
      "description": "Moved Permanently",
      "schema": {
       "$ref": "#/definitions/v1.Status"
      }
     },
     "401": {
      "description": "Unauthorized",
      "schema": {
       "$ref": "#/definitions/v1.Status"
      }
     },
     "500": {
      "description": "Internal server error",
      "schema": {
       "$ref": "#/definitions/v1.Status"
      }
     },
     "502": {
      "description": "Bad gateway",
      "schema": {
       "$ref": "#/definitions/v1.Status"
      }
     }
    }
   },
   "post": {
    "consumes": [
     "application/json"
    ],
    "produces": [
     "application/json"
    ],
    "summary": "Create organization",
    "operationId": "createOrganization",
    "parameters": [
     {
      "name": "body",
      "in": "body",
      "required": true,
      "schema": {
       "$ref": "#/definitions/v1alpha1.Organization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "OK",
      "schema": {
       "$ref": "#/definitions/v1alpha1.Organization"
      }
     },
     "301": {
      "description": "Moved Permanently",
      "schema": {
       "$ref": "#/definitions/v1.Status"
      }
     },
     "401": {
      "description": "Unauthorized",
      "schema": {
       "$ref": "#/definitions/v1.Status"
      }
     },
     "500": {
      "description": "Internal server error",
      "schema": {
       "$ref": "#/definitions/v1.Status"
      }
     },
     "502": {
      "description": "Bad gateway",
      "schema": {
       "$ref": "#/definitions/v1.Status"
      }
     }
    }
   }
  },
  "/faros.sh/api/v1alpha1/organizations/{organization}": {
   "get": {
    "consumes": [
     "application/json"
    ],
    "produces": [
     "application/json"
    ],
    "summary": "Get organization",
    "operationId": "getOrganization",
    "parameters": [
     {
      "type": "string",
      "description": "Name of an organization",
      "name": "organization",
      "in": "path",
      "required": true
     }
    ],
    "responses": {
     "200": {
      "description": "OK",
      "schema": {
       "$ref": "#/definitions/v1alpha1.Organization"
      }
     },
     "301": {
      "description": "Moved Permanently",
      "schema": {
       "$ref": "#/definitions/v1.Status"
      }
     },
     "401": {
      "description": "Unauthorized",
      "schema": {
       "$ref": "#/definitions/v1.Status"
      }
     },
     "500": {
      "description": "Internal server error",
      "schema": {
       "$ref": "#/definitions/v1.Status"
      }
     },
     "502": {
      "description": "Bad gateway",
      "schema": {
       "$ref": "#/definitions/v1.Status"
      }
     }
    }
   },
   "delete": {
    "consumes": [
     "application/json"
    ],
    "produces": [
     "application/json"
    ],
    "summary": "Delete organization",
    "operationId": "deleteOrganization",
    "parameters": [
     {
      "type": "string",
      "description": "Name of an organization",
      "name": "organization",
      "in": "path",
      "required": true
     }
    ],
    "responses": {
     "200": {
      "description": "OK",
      "schema": {
       "$ref": "#/definitions/v1.Status"
      }
     },
     "301": {
      "description": "Moved Permanently",
      "schema": {
       "$ref": "#/definitions/v1.Status"
      }
     },
     "401": {
      "description": "Unauthorized",
      "schema": {
       "$ref": "#/definitions/v1.Status"
      }
     },
     "500": {
      "description": "Internal server error",
      "schema": {
       "$ref": "#/definitions/v1.Status"
      }
     },
     "502": {
      "description": "Bad gateway",
      "schema": {
       "$ref": "#/definitions/v1.Status"
      }
     }
    }
   }
  },
  "/faros.sh/api/v1alpha1/organizations/{organization}/workspaces": {
   "get": {
    "consumes": [
     "application/json"
    ],
    "produces": [
     "application/json"
    ],
    "summary": "List workspaces",
    "operationId": "listWorkspaces",
    "parameters": [
     {
      "type": "string",
      "description": "Name of an organization",
      "name": "organization",
      "in": "path",
      "required": true
     }
    ],
    "responses": {
     "200": {
      "description": "OK",
      "schema": {
       "$ref": "#/definitions/v1alpha1.WorkspaceList"
      }
     },
     "301": {
      "description": "Moved Permanently",
      "schema": {
       "$ref": "#/definitions/v1.Status"
      }
     },
     "401": {
      "description": "Unauthorized",
      "schema": {
       "$ref": "#/definitions/v1.Status"
      }
     },
     "500": {
      "description": "Internal server error",
      "schema": {
       "$ref": "#/definitions/v1.Status"
      }
     },
     "502": {
      "description": "Bad gateway",
      "schema": {
       "$ref": "#/definitions/v1.Status"
      }
     }
    }
   },
   "post": {
    "consumes": [
     "application/json"
    ],
    "produces": [
     "application/json"
    ],
    "summary": "Create workspace",
    "operationId": "createWorkspace",
    "parameters": [
     {
      "name": "body",
      "in": "body",
      "required": true,
      "schema": {
       "$ref": "#/definitions/v1alpha1.Workspace"
      }
     },
     {
      "type": "string",
      "description": "Name of an organization",
      "name": "organization",
      "in": "path",
      "required": true
     }
    ],
    "responses": {
     "200": {
      "description": "OK",
      "schema": {
       "$ref": "#/definitions/v1alpha1.Organization"
      }
     },
     "301": {
      "description": "Moved Permanently",
      "schema": {
       "$ref": "#/definitions/v1.Status"
      }
     },
     "401": {
      "description": "Unauthorized",
      "schema": {
       "$ref": "#/definitions/v1.Status"
      }
     },
     "500": {
      "description": "Internal server error",
      "schema": {
       "$ref": "#/definitions/v1.Status"
      }
     },
     "502": {
      "description": "Bad gateway",
      "schema": {
       "$ref": "#/definitions/v1.Status"
      }
     }
    }
   }
  },
  "/faros.sh/api/v1alpha1/organizations/{organization}/workspaces/{workspace}": {
   "get": {
    "consumes": [
     "application/json"
    ],
    "produces": [
     "application/json"
    ],
    "summary": "Get workspace",
    "operationId": "getWorkspace",
    "parameters": [
     {
      "type": "string",
      "description": "Name of an workspace",
      "name": "workspace",
      "in": "path",
      "required": true
     },
     {
      "type": "string",
      "description": "Name of an organization",
      "name": "organization",
      "in": "path",
      "required": true
     }
    ],
    "responses": {
     "200": {
      "description": "OK",
      "schema": {
       "$ref": "#/definitions/v1alpha1.Workspace"
      }
     },
     "301": {
      "description": "Moved Permanently",
      "schema": {
       "$ref": "#/definitions/v1.Status"
      }
     },
     "401": {
      "description": "Unauthorized",
      "schema": {
       "$ref": "#/definitions/v1.Status"
      }
     },
     "500": {
      "description": "Internal server error",
      "schema": {
       "$ref": "#/definitions/v1.Status"
      }
     },
     "502": {
      "description": "Bad gateway",
      "schema": {
       "$ref": "#/definitions/v1.Status"
      }
     }
    }
   },
   "delete": {
    "consumes": [
     "application/json"
    ],
    "produces": [
     "application/json"
    ],
    "summary": "Delete workspace",
    "operationId": "deleteWorkspace",
    "parameters": [
     {
      "type": "string",
      "description": "Name of an workspace",
      "name": "workspace",
      "in": "path",
      "required": true
     },
     {
      "type": "string",
      "description": "Name of an organization",
      "name": "organization",
      "in": "path",
      "required": true
     }
    ],
    "responses": {
     "200": {
      "description": "OK",
      "schema": {
       "$ref": "#/definitions/v1.Status"
      }
     },
     "301": {
      "description": "Moved Permanently",
      "schema": {
       "$ref": "#/definitions/v1.Status"
      }
     },
     "401": {
      "description": "Unauthorized",
      "schema": {
       "$ref": "#/definitions/v1.Status"
      }
     },
     "500": {
      "description": "Internal server error",
      "schema": {
       "$ref": "#/definitions/v1.Status"
      }
     },
     "502": {
      "description": "Bad gateway",
      "schema": {
       "$ref": "#/definitions/v1.Status"
      }
     }
    }
   }
  }
 },
 "definitions": {
...
  "v1.ListMeta": {
   "description": "ListMeta describes metadata that synthetic resources must have, including lists and various status objects. A resource may have only one of {ObjectMeta, ListMeta}.",
   "properties": {
    "continue": {
     "description": "continue may be set if the user set a limit on the number of items returned, and indicates that the server has more data available. The value is opaque and may be used to issue another request to the endpoint that served this list to retrieve the next set of available objects. Continuing a consistent list may not be possible if the server configuration has changed or more than a few minutes have passed. The resourceVersion field returned when using this continue value will be identical to the value in the first response, unless you have received this token from an error message.",
     "type": "string"
    },
    "remainingItemCount": {
     "description": "remainingItemCount is the number of subsequent items in the list which are not included in this list response. If the list request contained label or field selectors, then the number of remaining items is unknown and the field will be left unset and omitted during serialization. If the list is complete (either because it is not chunking or because this is the last chunk), then there are no more remaining items and this field will be left unset and omitted during serialization. Servers older than v1.15 do not set this field. The intended use of the remainingItemCount is *estimating* the size of a collection. Clients should not rely on the remainingItemCount to be set or to be exact.",
     "type": "integer",
     "format": "int64"
    },
    "resourceVersion": {
     "description": "String that identifies the server's internal version of this object that can be used by clients to determine when objects have changed. Value must be treated as opaque by clients and passed unmodified back to the server. Populated by the system. Read-only. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#concurrency-control-and-consistency",
     "type": "string"
    },
    "selfLink": {
     "description": "Deprecated: selfLink is a legacy read-only field that is no longer populated by the system.",
     "type": "string"
    }
   }
  },
...
  "v1.ObjectReference": {
   "description": "ObjectReference contains enough information to let you inspect or modify the referred object.",
   "properties": {
    "apiVersion": {
     "description": "API version of the referent.",
     "type": "string"
    },
    "fieldPath": {
     "description": "If referring to a piece of an object instead of an entire object, this string should contain a valid JSON/Go field access statement, such as desiredState.manifest.containers[2]. For example, if the object reference is to a container within a pod, this would take on a value like: \"spec.containers{name}\" (where \"name\" refers to the name of the container that triggered the event) or if no container name is specified \"spec.containers[2]\" (container with index 2 in this pod). This syntax is chosen only to have some well-defined way of referencing a part of an object.",
     "type": "string"
    },
    "kind": {
     "description": "Kind of the referent. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds",
     "type": "string"
    },
    "name": {
     "description": "Name of the referent. More info: https://kubernetes.io/docs/concepts/overview/working-with-objects/names/#names",
     "type": "string"
    },
    "namespace": {
     "description": "Namespace of the referent. More info: https://kubernetes.io/docs/concepts/overview/working-with-objects/namespaces/",
     "type": "string"
    },
    "resourceVersion": {
     "description": "Specific resourceVersion to which this reference is made, if any. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#concurrency-control-and-consistency",
     "type": "string"
    },
    "uid": {
     "description": "UID of the referent. More info: https://kubernetes.io/docs/concepts/overview/working-with-objects/names/#uids",
     "type": "string"
    }
   }
  },
...
  "v1.Status": {
   "description": "Status is a return value for calls that don't return other objects.",
   "properties": {
    "apiVersion": {
     "description": "APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources",
     "type": "string"
    },
    "code": {
     "description": "Suggested HTTP return code for this status, 0 if not set.",
     "type": "integer",
     "format": "int32"
    },
    "details": {
     "description": "Extended data associated with the reason.  Each reason may define its own extended details. This field is optional and the data returned is not guaranteed to conform to any schema except that defined by the reason type.",
     "$ref": "#/definitions/v1.StatusDetails"
    },
    "kind": {
     "description": "Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds",
     "type": "string"
    },
    "message": {
     "description": "A human-readable description of the status of this operation.",
     "type": "string"
    },
    "metadata": {
     "description": "Standard list metadata. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds",
     "$ref": "#/definitions/v1.ListMeta"
    },
    "reason": {
     "description": "A machine-readable description of why this operation is in the \"Failure\" status. If this value is empty there is no information available. A Reason clarifies an HTTP status code but does not override it.",
     "type": "string"
    },
    "status": {
     "description": "Status of the operation. One of: \"Success\" or \"Failure\". More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status",
     "type": "string"
    }
   }
  },
  "v1.StatusCause": {
   "description": "StatusCause provides more information about an api.Status failure, including cases when multiple errors are encountered.",
   "properties": {
    "field": {
     "description": "The field of the resource that has caused this error, as named by its JSON serialization. May include dot and postfix notation for nested attributes. Arrays are zero-indexed.  Fields may appear more than once in an array of causes due to fields having multiple errors. Optional.\n\nExamples:\n  \"name\" - the field \"name\" on the current resource\n  \"items[0].name\" - the field \"name\" on the first array entry in \"items\"",
     "type": "string"
    },
    "message": {
     "description": "A human-readable description of the cause of the error.  This field may be presented as-is to a reader.",
     "type": "string"
    },
    "reason": {
     "description": "A machine-readable description of the cause of the error. If this value is empty there is no information available.",
     "type": "string"
    }
   }
  },
  "v1.StatusDetails": {
   "description": "StatusDetails is a set of additional properties that MAY be set by the server to provide additional information about a response. The Reason field of a Status object defines what attributes will be set. Clients must ignore fields that do not match the defined type of each attribute, and should assume that any attribute may be empty, invalid, or under defined.",
   "properties": {
    "causes": {
     "description": "The Causes array includes more details associated with the StatusReason failure. Not all StatusReasons may provide detailed causes.",
     "type": "array",
     "items": {
      "$ref": "#/definitions/v1.StatusCause"
     }
    },
    "group": {
     "description": "The group attribute of the resource associated with the status StatusReason.",
     "type": "string"
    },
    "kind": {
     "description": "The kind attribute of the resource associated with the status StatusReason. On some operations may differ from the requested resource Kind. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds",
     "type": "string"
    },
    "name": {
     "description": "The name attribute of the resource associated with the status StatusReason (when there is a single name which can be described).",
     "type": "string"
    },
    "retryAfterSeconds": {
     "description": "If specified, the time in seconds before the operation should be retried. Some errors may indicate the client must take an alternate action - for those errors this field may indicate how long to wait before taking the alternate action.",
     "type": "integer",
     "format": "int32"
    },
    "uid": {
     "description": "UID of the resource. (when there is a single resource which can be described). More info: http://kubernetes.io/docs/user-guide/identifiers#uids",
     "type": "string"
    }
   }
  },
...
  "v1alpha1.Condition": {
   "required": [
    "type",
    "status",
    "lastTransitionTime"
   ],
   "properties": {
    "lastTransitionTime": {
     "type": "string"
    },
    "message": {
     "type": "string"
    },
    "reason": {
     "type": "string"
    },
    "severity": {
     "type": "string"
    },
    "status": {
     "type": "string"
    },
    "type": {
     "type": "string"
    }
   }
  },
  "v1alpha1.ObjectReference": {
   "properties": {
    "apiVersion": {
     "type": "string"
    },
    "email": {
     "type": "string"
    },
    "kind": {
     "type": "string"
    },
    "name": {
     "type": "string"
    }
   }
  },
  "v1alpha1.Organization": {
   "properties": {
    "apiVersion": {
     "description": "APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources",
     "type": "string"
    },
    "kind": {
     "description": "Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds",
     "type": "string"
    },
    "metadata": {
     "$ref": "#/definitions/v1.ObjectMeta"
    },
    "spec": {
     "$ref": "#/definitions/v1alpha1.OrganizationSpec"
    },
    "status": {
     "$ref": "#/definitions/v1alpha1.OrganizationStatus"
    }
   }
  },
  "v1alpha1.OrganizationList": {
   "required": [
    "items"
   ],
   "properties": {
    "apiVersion": {
     "description": "APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources",
     "type": "string"
    },
    "items": {
     "type": "array",
     "items": {
      "$ref": "#/definitions/v1alpha1.Organization"
     }
    },
    "kind": {
     "description": "Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds",
     "type": "string"
    },
    "metadata": {
     "$ref": "#/definitions/v1.ListMeta"
    }
   }
  },
  "v1alpha1.OrganizationSpec": {
   "properties": {
    "description": {
     "type": "string"
    },
    "ownersRef": {
     "type": "array",
     "items": {
      "$ref": "#/definitions/v1alpha1.ObjectReference"
     }
    }
   }
  },
  "v1alpha1.OrganizationStatus": {
   "properties": {
    "conditions": {
     "type": "array",
     "items": {
      "$ref": "#/definitions/v1alpha1.Condition"
     }
    },
    "workspaceURL": {
     "type": "string"
    }
   }
  },
  "v1alpha1.Workspace": {
   "properties": {
    "apiVersion": {
     "description": "APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources",
     "type": "string"
    },
    "kind": {
     "description": "Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds",
     "type": "string"
    },
    "metadata": {
     "$ref": "#/definitions/v1.ObjectMeta"
    },
    "spec": {
     "$ref": "#/definitions/v1alpha1.WorkspaceSpec"
    },
    "status": {
     "$ref": "#/definitions/v1alpha1.WorkspaceStatus"
    }
   }
  },
  "v1alpha1.WorkspaceList": {
   "required": [
    "items"
   ],
   "properties": {
    "apiVersion": {
     "description": "APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources",
     "type": "string"
    },
    "items": {
     "type": "array",
     "items": {
      "$ref": "#/definitions/v1alpha1.Workspace"
     }
    },
    "kind": {
     "description": "Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds",
     "type": "string"
    },
    "metadata": {
     "$ref": "#/definitions/v1.ListMeta"
    }
   }
  },
  "v1alpha1.WorkspaceSpec": {
   "properties": {
    "description": {
     "type": "string"
    },
    "organizationRef": {
     "$ref": "#/definitions/v1.ObjectReference"
    }
   }
  },
  "v1alpha1.WorkspaceStatus": {
   "properties": {
    "conditions": {
     "type": "array",
     "items": {
      "$ref": "#/definitions/v1alpha1.Condition"
     }
    },
    "workspaceURL": {
     "type": "string"
    }
   }
  }
 },
 "securityDefinitions": {
  "OAuth2": {
   "type": "oauth2",
   "flow": "accessCode",
   "authorizationUrl": ""
  }
 },
 "security": [
  {
   "OAuth2": []
  }
 ],
...
}